DAO-203 Rename pages/proposal-commons -> pages/components
