+++ src/app.tsx
@@ -13,7 +13,7 @@ import Dashboard from './pages/dashboard';
 import History from './pages/history';
 import HistoryDetails from './pages/history-details';
 import NotFoundPage from './pages/not-found';
-import ProposalDetailsPage from './pages/proposal-commons/proposal-details';
+import ProposalDetailsPage from './pages/components/proposal-details';
 import Proposals from './pages/proposals';
 import Vesting from './pages/vesting';
 import Policies from './pages/my-policies';
